+++ Implementations/BuildingPerTheme/Math/CheckerboardFloor.shader
@@ -22,9 +22,9 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/Math/CheckerboardFloor"
 
             float _zoom;
             float _lineThickness;
-            half4 _lineColor;
-            half4 _colorA;
-            half4 _colorB;
+            float4 _lineColor;
+            float4 _colorA;
+            float4 _colorB;
 
             struct appdata {
                 float4 vertex : POSITION;
@@ -46,7 +46,7 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/Math/CheckerboardFloor"
                 return o;
             }
 
-            half4 frag (v2f i) : SV_Target {
+            float4 frag (v2f i) : SV_Target {
                 float3 worldNormal = normalize(i.worldNormal);
 
                 float2 pos2D = i.vertex.xy;
@@ -60,20 +60,20 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/Math/CheckerboardFloor"
                 float2 linePattern2 = step(1.0 - _lineThickness, sawtooth2);
                 float linePattern = saturate(linePattern2.x + linePattern2.y);
 
-                half3 albedo = lerp(_colorA, _colorB, checkerboardPattern);
+                float3 albedo = lerp(_colorA, _colorB, checkerboardPattern);
                 albedo = lerp(albedo, _lineColor, linePattern);
 
                 // Light
                 float3 _world_light_direction = normalize(_WorldSpaceLightPos0.xyz);
                 float3 lightColor = _LightColor0.rgb;
-                half3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
-                half3 NdotL = dot(worldNormal, _world_light_direction);
+                float3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
+                float3 NdotL = dot(worldNormal, _world_light_direction);
                 NdotL = saturate(NdotL);
-                half3 directLight = NdotL * lightColor.rgb;
+                float3 directLight = NdotL * lightColor.rgb;
 
-                half3 color = half3((directLight + ambientLight) * albedo);
+                float3 color = float3((directLight + ambientLight) * albedo);
 
-                return half4 (color, 1.0);
+                return float4 (color, 1.0);
             }
             ENDCG
         }
