+++ Implementations/Environment/Terrain/TriplanarTerrain.shader
@@ -49,7 +49,7 @@ Shader "ShaderCastle/Implementations/Environment/TriplanarTerrain"
                 return o;
             }
 
-            half3 triplanarColor(sampler2D lookupTexture, float3 pos, float3 normal) {
+            float3 triplanarColor(sampler2D lookupTexture, float3 pos, float3 normal) {
                 float3 weights = abs(normal);
                 weights = pow(weights, _Sharpness);
                 weights /= (weights.x + weights.y + weights.z);
@@ -60,32 +60,32 @@ Shader "ShaderCastle/Implementations/Environment/TriplanarTerrain"
                 float2 uvY = float2(pos.x * direction.y, pos.z) * _MainTexScale;
                 float2 uvZ = float2(pos.x * -direction.z, pos.y) * _MainTexScale;
 
-                half4 colX = tex2D(lookupTexture, uvX);
-                half4 colY = tex2D(lookupTexture, uvY);
-                half4 colZ = tex2D(lookupTexture, uvZ);
+                float4 colX = tex2D(lookupTexture, uvX);
+                float4 colY = tex2D(lookupTexture, uvY);
+                float4 colZ = tex2D(lookupTexture, uvZ);
 
-                half4 finalCol = colX * weights.x + colY * weights.y + colZ * weights.z;
+                float4 finalCol = colX * weights.x + colY * weights.y + colZ * weights.z;
                 return finalCol;
             }
 
-            half4 frag (v2f i) : SV_Target {
+            float4 frag (v2f i) : SV_Target {
                 float3 worldNormal = normalize(i.worldNormal);
 
-                half3 cliff = triplanarColor(_CliffTexture, i.localPos, i.normal);
-                half3 top = tex2D(_TopTexture, i.localPos.xz);
+                float3 cliff = triplanarColor(_CliffTexture, i.localPos, i.normal);
+                float3 top = tex2D(_TopTexture, i.localPos.xz);
                 float mask = smoothstep(_Transition - 0.05, _Transition + 0.05, i.normal.y);
                 float3 albedo = lerp(cliff, top, mask);
 
                 float3 _world_light_direction = normalize(_WorldSpaceLightPos0.xyz);
                 float3 lightColor = _LightColor0.rgb;
-                half3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
-                half3 NdotL = dot(worldNormal, _world_light_direction);
+                float3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
+                float3 NdotL = dot(worldNormal, _world_light_direction);
                 NdotL = saturate(NdotL);
-                half3 directLight = NdotL * lightColor.rgb;
+                float3 directLight = NdotL * lightColor.rgb;
 
-                half3 color = half3((directLight + ambientLight) * albedo);
+                float3 color = float3((directLight + ambientLight) * albedo);
 
-                return half4 (color, 1.0);
+                return float4 (color, 1.0);
             }
             ENDCG
         }
